fix: broken tests
