+++ src/tests/Feature/Controller/PostControllerTest.php
@@ -240,54 +240,17 @@ class PostControllerTest extends TestCase
     }
 
     /**
-     * Tests that we return the images page and data correctly.
+     * Tests that we generate the reading time correctly.
      *
      * @return void
      */
-    public function testShowImages()
+    public function testGenerateReadingTime()
     {
         $controller = new PostController();
+        $post = Post::factory()->create();
 
-        $request = Request::create('/images', 'GET');
-        $controller->showImagesPage($request);
+        $readingTime = $controller::generateReadingTime($post);
 
-        // TODO: Is there a better assertion here since we return an empty view?
-        $this->assertTrue(true);
-    }
-
-    /**
-     * Tests that we can upload a post image correctly.
-     *
-     * @return void
-     */
-    public function testUpdateProfilePic()
-    {
-        // TODO: Finish writing this test asserting an image got uploaded
-        $this->doesNotPerformAssertions();
-        // $controller = new PostController();
-
-        // $request = Request::create("/images", 'POST');
-        // $response = $controller->uploadPostImage($request);
-
-        // $this->assertEquals('Image uploaded successfully.', $response->getSession()->get('message'));
-        // $this->assertEquals(302, $response->getStatusCode());
-    }
-
-    /**
-     * Tests that we can delete a post image correctly.
-     *
-     * @return void
-     */
-    public function testDeletePostImage()
-    {
-        // TODO: Finish writing this test asserting an image got uploaded
-        $this->doesNotPerformAssertions();
-        // $controller = new PostController();
-
-        // $request = Request::create("/images", 'DELETE');
-        // $response = $controller->deletePostImage($request);
-
-        // $this->assertEquals('Image deleted.', $response->getSession()->get('message'));
-        // $this->assertEquals(302, $response->getStatusCode());
+        $this->assertEquals(1, $readingTime);
     }
 }
